Refactoring no metodo entropy em Utils e refactoring da classe Classifier com revisao de exceptions e valores Double.

+++ Datamining/src/datamining/Utils.java
@@ -4,7 +4,6 @@
  */
 package datamining;
 
-import java.util.LinkedList;
 import static java.lang.Math.*;
 
 /**
@@ -46,17 +45,18 @@ public class Utils {
      * 
      * @return o valor da entropia para as probabilidades passadas
      */
-    public static Double entropy(Double[] probs) {
+    public static double entropy(double[] probs) {
         double entropyValue = 0;
+        
         for (int i = 0; i < probs.length; i++) {
-            double px = probs[i].doubleValue();
+            double px = probs[i];
             double log2 = 0;
             if (px != 0) {
                 log2 = log(px) / log(2);
             }
             entropyValue += px * log2;
         }
-        return Double.valueOf(-entropyValue);
+        return -entropyValue;
     }
 
 }
